debug: fix the issue of memory out of boundary
When run the example directly, it fails. When run it with valgrind, it
could pass.

It's caused by struct wd_state is attached at the end of struct
internal_state. But the allocation function doesn't follow the rule.
It causes accessing memory out of boundary. Now fix it in
wd_alloc_state().

But we meet a new issue. If run it directly, it triggers an exception
in deflateInit2_() when execute "s->strm = strm;". If run it with
valgrind, it could pass. Before "s" is initialized with allocated
memory address, it's 0xffffbf6a50d8. After "s" is initialized with
allocated memory address, it's 0xffffffffaaae58d0. Apparently, the
address value is invalid that causes an exception.

Signed-off-by: Haojian Zhuang <haojian.zhuang@linaro.org>