debug: fix the issue of invalid memory address
If run it directly, it triggers an exception in deflateInit2_() when
execute "s->strm = strm;". If run it with valgrind, it could pass.
Before "s" is initialized with allocated memory address, it's 0xffffbf6a50d8.
After "s" is initialized with allocated memory address, it's
0xffffffffaaae58d0. Apparently, the address value is invalid that causes
an exception.

Actually, "wd_common.h" is missed in "wd_deflate.h". So ZALLOC may be
linked wrongly.

Signed-off-by: Haojian Zhuang <haojian.zhuang@linaro.org>